third_party/blink/web_tests/transitions/transition-end-event-prefixed-01-expected.txt

Test to make sure that if prefixed transition events are modified we correctly modify unprefixed events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS transitionEventContainer.type is transitionEventBox.type
PASS transitionEventContainer.bubbles is transitionEventBox.bubbles
PASS transitionEventContainer.timeStamp is transitionEventBox.timeStamp
PASS transitionEventContainer.cancelable is transitionEventBox.cancelable
PASS transitionEventContainer.srcElement is transitionEventBox.srcElement
PASS transitionEventContainer.cancelBubble is transitionEventBox.cancelBubble
PASS transitionEventContainer.defaultPrevented is transitionEventBox.defaultPrevented
PASS transitionEventContainer.target is transitionEventBox.target
PASS transitionEventContainer.currentTarget is testContainer
PASS transitionEventContainer.pseudoElement is transitionEventBox.pseudoElement
PASS transitionEventContainer.elapsedTime is transitionEventBox.elapsedTime
PASS transitionEventContainer.propertyName is transitionEventBox.propertyName
